wbsmicroservice(backend) da testare!

+++ backend/WBSMicroservice/src/main/java/com/WBSMicroservice/controller/InstructionController.java
@@ -1,6 +1,9 @@
 package com.WBSMicroservice.controller;
 
+import java.io.UnsupportedEncodingException;
+import java.util.LinkedHashMap;
 import java.util.List;
+import java.util.Map;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -15,32 +18,75 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.WBSMicroservice.dto.InstructionDTO;
+import com.WBSMicroservice.dto.ParamDTO;
 import com.WBSMicroservice.services.InstructionService;
+import com.WBSMicroservice.utils.JwtUtils;
 
-@CrossOrigin(value="*")
+import io.jsonwebtoken.ExpiredJwtException;
+
+@CrossOrigin(value = "*")
 @RestController
 @RequestMapping("/Instruction")
 public class InstructionController {
-	
+
 	@Autowired
 	private InstructionService istruzioneService;
-	
-	public InstructionController() {}
-	
+
+	public InstructionController() {
+	}
+
 	@PostMapping("/insertInstruction")
-	public ResponseEntity<InstructionDTO> insertInstruction(@RequestBody InstructionDTO instruction) {
-		//System.out.println("Instruction: "+instruction);
-		return ResponseEntity.status(HttpStatus.OK).body(istruzioneService.insertIstruzione(instruction));
+	public ResponseEntity<InstructionDTO> insertInstruction(@RequestBody ParamDTO param) {
+		LinkedHashMap instruction = (LinkedHashMap) param.getParam();
+		int rank;
+		try {
+			rank = this.getRankFromJwt(param.getJwt());
+			if (rank == 0)
+				return ResponseEntity.status(HttpStatus.OK)
+						.body(istruzioneService.insertIstruzione((InstructionDTO) instruction.get(0)));
+			else
+				return ResponseEntity.status(HttpStatus.FORBIDDEN).body(null);
+		} catch (ExpiredJwtException | UnsupportedEncodingException e) {
+			return ResponseEntity.status(HttpStatus.FORBIDDEN).body(null);
+		}
 	}
-	
+
 	@DeleteMapping("/deleteInstruction")
-	public ResponseEntity<Boolean>  deleteInstruction(@RequestParam(value="idInstruction") int idInstruction) {		
-		return ResponseEntity.status(HttpStatus.OK).body(istruzioneService.deleteIstruzione(idInstruction));
+	public ResponseEntity<Boolean> deleteInstruction(@RequestParam(value = "jwt") String jwt,
+			@RequestParam(value = "idInstruction") int idInstruction) {
+		int rank;
+		try {
+			rank = this.getRankFromJwt(jwt);
+			if (rank == 0)
+				return ResponseEntity.status(HttpStatus.OK).body(istruzioneService.deleteIstruzione(idInstruction));
+			else
+				return ResponseEntity.status(HttpStatus.FORBIDDEN).body(null);
+		} catch (ExpiredJwtException | UnsupportedEncodingException e) {
+			return ResponseEntity.status(HttpStatus.FORBIDDEN).body(null);
+		}
 	}
-	
+
 	@GetMapping("/showInstruction")
-	public ResponseEntity<List<InstructionDTO>> showInstruction(@RequestParam(value="idTask") int idTask) {	
-		return ResponseEntity.status(HttpStatus.OK).body(istruzioneService.getAllIstruzioniByIdTask(idTask));
+	public ResponseEntity<List<InstructionDTO>> showInstruction(@RequestParam(value = "jwt") String jwt,
+			@RequestParam(value = "idTask") int idTask) {
+		try {
+			int rank = this.getRankFromJwt(jwt);
+			if (rank == 0)
+				return ResponseEntity.status(HttpStatus.OK).body(istruzioneService.getAllIstruzioniByIdTask(idTask));
+			else
+				return ResponseEntity.status(HttpStatus.FORBIDDEN).body(null);
+		} catch (ExpiredJwtException | UnsupportedEncodingException e) {
+			return ResponseEntity.status(HttpStatus.FORBIDDEN).body(null);
+		}
+
+	}
+
+	private int getRankFromJwt(String jwt) throws ExpiredJwtException, UnsupportedEncodingException {
+
+		Map<String, Object> data = JwtUtils.jwt2Map(jwt);
+		int rank = Integer.parseInt(data.get("scope").toString());
+
+		return rank;
 	}
 
 }
